feat: using argparse to create a CLI:
     this feature will make user to use the script from command line
with some arguments that specified by the user.

with this aproach there is no need to prompt user for any input. script will
use the config.ini if user does not pass any args, or user can choose
other path for config.ini or pass each args with coresponding switch.

if any switch does not pass to the script, script will use the default
value in define in the script it self.